requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
google-auth==2.40.2
greenlet==3.2.1
groq==0.24.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-groq==0.3.2
langchain-huggingface==0.2.0
langchain-pinecone==0.2.6
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
markupsafe==3.0.2
marshmallow==3.26.1
motor==3.7.1
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pillow==11.2.1
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
propcache==0.3.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pymongo==4.13.0
pypdf==5.4.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==80.3.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.2
sympy==1.14.0
syrupy==4.9.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
